refactor radar plot controls
